# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OxyGen CMS API
  version: 0.8.0
  summary: OxyGen Central Management Server API
  description: |
    API contract for OxyGen Central Management Server Phase 1.

    Current CMS schema version: 0.19.

    Phase 1 APIs cover first-run setup, local authentication/RBAC, tenant/user/group/role administration,
    manual OxyGen instance enrollment, dashboard background poller status/controls, application logs, grid preferences, application settings,
    Settings system visibility, and Phase 1.5 queue status. Protected routes use bearer authentication plus explicit permissions and Tenant/global/instance scope checks.
servers:
  - url: http://localhost:3000
    description: Local API development server
  - url: https://{host}
    description: Deployed CMS server
    variables:
      host:
        default: cms.example.com
security:
  - bearerAuth: []
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Sign in
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Bearer token and user profile.
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    required: [token]
                    properties:
                      token: { type: string }
                  - $ref: '#/components/schemas/AuthProfile'
        '401': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
...
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        details: {}
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1, writeOnly: true }
...
    PublicUser:
      type: object
      required: [id, email, displayName, tenantId, instanceAccessMode, instanceIds, isActive, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        displayName: { type: string }
        tenantId: { type: [string, 'null'], format: uuid }
        instanceAccessMode: { $ref: '#/components/schemas/UserInstanceAccessMode' }
        instanceIds:
          type: array
          items: { type: string, format: uuid }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AuthProfile:
      type: object
      required: [user, roles, permissions, groups]
      properties:
        user: { $ref: '#/components/schemas/PublicUser' }
        roles:
          type: array
          items: { type: string }
        permissions:
          type: array
          items: { type: string }
          description: Effective permission keys resolved from direct role assignments and group membership; the UI uses these for capability gates, but backend route authorization remains authoritative.
        groups:
          type: array
          items:
            type: object
            required: [id, name, tenantId, instanceAccessMode, instanceIds]
            properties:
              id: { type: string, format: uuid }
              name: { type: string }
              tenantId: { type: [string, 'null'], format: uuid }
              instanceAccessMode: { $ref: '#/components/schemas/GroupInstanceAccessMode' }
              instanceIds:
                type: array
                items: { type: string, format: uuid }
...
    GroupInstanceAccessMode:
      type: string
      enum: [none, all, specific]
      default: none
    UserInstanceAccessMode:
      type: string
      enum: [inherit, none, all, specific]
      default: inherit